本体と共有ライブラリを別に生成できるように分割
