+++ src/almo.cpp
@@ -5,11 +5,6 @@
 #include <string>
 #include "utils.hpp"
 
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h> 
-
-namespace py = pybind11;
-
 class Config {
 public:
     std::string template_file = "__default__";
@@ -179,14 +174,3 @@ int main(int argc, char* argv[]) {
     return 0;
 }
 
-std::string md_to_html(std::string md_content, std::map<std::string, std::string> meta_data) {
-    almo::Block ast = almo::parse_rest(split(md_content, "\n"));
-    return ast.render(meta_data);
-}
-
-
-PYBIND11_MODULE(almo, m) {
-    m.doc() = "almo library"; 
-
-    m.def("md_to_html", &md_to_html, "md to html");
-}
